Consider all three cases of loading from cache
From the last to the first as in code.
First case: we load old Articles, that definetely stored in database.
Second case: we load top Articles. If they are loaded, give them as callback.
If database is empty, load Articles first, then give them.
Third case: we load new Articles. If they are loaded, give them. If there is
not enough new Articles to display, firstly load new from server.