Add 'last' persistancy via SharedPreferences
As user may have not seen the most recent posts yet and moreover his current
top post may be far from the most recent loaded posts (the distance
theoretically can be up to ~100 posts), it is very helpful to UX to keep this
index of the top post in View. Actually, the 'last' already serves to this
mission during one application session. To maintain it through different
launches, it was put in SharedPreferences.