Implement Style class
This class defines some style. Style is a rule of what should be replaced with
what. All styles should implement this class by providing original regular
expression which should be replaced and what will be instead. As an example, to
create style that removes all 'a' letters, getOriginal() should return String
"a" and getReplacement() should return an empty String.